+++ packages/locadex/CHANGELOG.md
@@ -7,6 +7,20 @@
 - Updated dependencies [01ba884]
   - gt@2.16.3-iris.0
 
+## 1.0.207
+
+### Patch Changes
+
+- Updated dependencies []:
+  - gt@2.16.4
+
+## 1.0.206
+
+### Patch Changes
+
+- Updated dependencies [[`79e6836`](https://github.com/generaltranslation/gt/commit/79e6836349191220ee8f5848b5e6ff287246f162)]:
+  - gt@2.16.3
+
 ## 1.0.205
 
 ### Patch Changes