+++ packages/node/CHANGELOG.md
@@ -7,6 +7,14 @@
 - Updated dependencies [8d376e2]
   - gt-i18n@1.0.12-iris.0
 
+## 1.0.13
+
+### Patch Changes
+
+- Updated dependencies [[`b00b93e`](https://github.com/generaltranslation/gt/commit/b00b93eb3b830b8528ef3dbd5f503ff76d1b338a)]:
+  - generaltranslation@9.1.2
+  - gt-i18n@1.0.12
+
 ## 1.0.12
 
 ### Patch Changes